clean up tests, add some error handling

+++ db/real_fs.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"crypto/md5"
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -15,7 +16,7 @@ type RealFS struct {
 	RootDir string
 }
 
-func (fs RealFS) WritePhoto(photo Photo) {
+func (fs RealFS) WritePhoto(photo Photo) error {
 	photoFilepath := fs.PhotoFilepath(photo)
 	dirs := strings.Split(photoFilepath, "/")
 	photoDir := strings.Join(dirs[:len(dirs)-1], "/")
@@ -23,7 +24,12 @@ func (fs RealFS) WritePhoto(photo Photo) {
 	err := os.MkdirAll(photoDir, 0755)
 
 	if err != nil {
-		log.Fatal("Unable to create the photo dir:  ", err)
+		log.Print("Unable to create the photo dir:  ", err)
+		return err
+	}
+
+	if fs.photoAlreadyExists(photoFilepath) {
+		return errors.New("photo with that id already saved, something went wrong...")
 	}
 
 	out, err := os.Create(photoFilepath)
@@ -31,18 +37,30 @@ func (fs RealFS) WritePhoto(photo Photo) {
 
 	if err != nil {
 		log.Fatal("Unable to create the file for writing:  ", err)
+		return err
 	}
 
 	_, err = io.Copy(out, *photo.PhotoFile)
 	if err != nil {
 		log.Fatal("Unable to copy photo file for photo #", photo.Id, ":  ", err)
+		return err
 	}
+	return nil
+}
+
+func (fs RealFS) photoAlreadyExists(path string) bool {
+	_, err := os.Stat(path)
+
+	return !os.IsNotExist(err)
+
 }
 
 func (fs RealFS) CountPhotos() int {
 	return 50 // TODO count files in srv directory
 }
 
+// This might eventually go in the app package, since the presentation layer needs it too, but I'll
+// wait until I actually implement a new FS for that
 func (fs RealFS) PhotoSrc(id int) string {
 	paddedId := fmt.Sprintf("%012d", id)
 	imgMD5 := md5.Sum([]byte(paddedId))
